requirements.txt
Flask==3.1.1
flask-cors==6.0.1
python-dotenv==1.2.1
joblib==1.5.2
scikit-learn==1.7.2
requests==2.32.5
numpy==2.3.4
scipy==1.16.3
threadpoolctl==3.6.0
